Convert type to integer

+++ database/migrations/2020_10_03_210342_create_experiences_table.php
@@ -14,9 +14,10 @@ class CreateExperiencesTable extends Migration
     public function up()
     {
         Schema::create('experiences', function (Blueprint $table) {
+
             $table->id();
 
-            $table->tinyInteger('year');
+            $table->integer('year');
             $table->string('name_en', 100)->unique();
             $table->string('name_fr', 100)->unique();
             $table->string('description_en', 255)->nullable(true);
